+++ packages/relax-react/src/hooks/index.ts
@@ -3,8 +3,8 @@
  * Provides reactive integration between Relax state and React components
  */
 
-import type { State, Value } from '@relax-state/core';
-import { useCallback, useEffect, useState } from 'react';
+import type { Action, action, State, Value } from '@relax-state/core';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useRelaxStore } from '../provider';
 
 /**
@@ -54,4 +54,15 @@ export const useRelaxState = <T>(state: State<T>): readonly [T, (value: T) => vo
   );
 
   return [value, setState] as const;
+};
+
+export const useActions = <const P extends Action[]>(actions: P) => {
+  const store = useRelaxStore();
+  return useMemo(() => {
+    return actions.map(
+      (action) => (payload: Parameters<typeof action>[1]) => action(store, payload)
+    );
+  }, [actions, store]) as {
+    [K in keyof P]: P[K] extends Action<infer P, infer R> ? (payload: P) => R : never;
+  };
 };