+++ packages/relax-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @relax-state/react
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @relax-state/core@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes